HID: nintendo: Add support for the Nintendo Switch SNES controller.
It basically acts like a Pro controller but without analog sticks.
A separate mapping has been added to correctly represent its button
layout and report the d-pad inputs as a hat instead ofseparate buttons.

A macro to determine whether or not the controller has a USB port has also
been added to make the code clearer when this is checked since there are
now more possible devices.

Signed-off-by: Nadia Holmquist Pedersen <nadia@nhp.sh>